refactor(entity-list): clean up query view
Refs: TOCDEV-4806
